* Fixed bug in libhaggle that caused the id hash of a data object to become different depending on whether you run on 32bit or 64bit machines. * Update TestApp.java with example of using dispose() on data objects

+++ src/libhaggle/dataobject.c
@@ -527,7 +527,7 @@ int haggle_dataobject_calculate_id(const struct dataobject *dobj, dataobject_id_
 
 		list_for_each(pos, &dobj->al->attributes) {
 			haggle_attr_t *a = (haggle_attr_t *)pos;
-			unsigned long w;
+			u_int32_t w;
 
 			SHA1_Update(&ctxt, (unsigned char *)haggle_attribute_get_name(a), strlen(haggle_attribute_get_name(a)));
 			SHA1_Update(&ctxt, (unsigned char *)haggle_attribute_get_value(a), strlen(haggle_attribute_get_value(a)));
